requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
filter==0.0.0.20200724
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
